SmartHomeVisionSuite/requirements.txt
click==8.1.7
numpy==2.0.0
opencv-python==4.10.0.84
PyYAML==6.0.1
ffmpeg-python
requests